@@ -1,1 +1,1 @@
-@javax.xml.bind.annotation.XmlSchema(namespace = "http://ws.monitor.myaktion.dpunkt.de/") package de.dpunkt.myaktion.monitor.ws;
+@jakarta.xml.bind.annotation.XmlSchema(namespace = "http://ws.monitor.myaktion.dpunkt.de/") package de.dpunkt.myaktion.monitor.ws;
